% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{create_groups}
\alias{create_groups}
\title{Create roughly equal sized groups of people}
\usage{
create_groups(df_participants, min_group_size = 3)
}
\arguments{
\item{df_participants}{a data frame of willing participants!}

\item{min_group_size}{a number for the desired group size}
}
\value{
df.participants with a new column, Group
}
\description{
This function takes a data frame and creates smaller subsets
    based on the group.size parameter
}
