% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_topics.R
\name{create_topics}
\alias{create_topics}
\title{Create a list of topics}
\usage{
create_topics(group_nums, topic_vector)
}
\arguments{
\item{group_nums}{number of groups}

\item{topic_vector}{a vector of suggested topics}
}
\value{
a two-column data frame
}
\description{
This function returns a data frame including a group number and
    a suggested topic to kick-off the conversation.
}
