% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fika-package.R
\docType{package}
\name{fika-package}
\alias{fika}
\alias{fika-package}
\title{fika: Simple Tools for Organizing Fika Groups & Topics}
\description{
`fika` is a package is to help organize large groups into smaller Fikas &
    suggest topics of discussion.
}
\author{
\strong{Maintainer}: Matt Leary \email{matt.leary@protonmail.com}

}
\keyword{internal}
