% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_main.R
\name{open_main}
\alias{open_main}
\title{Open the run script to execute code}
\usage{
open_main()
}
\value{
nothing.  Open the main script
}
\description{
Open up the main script to execute the package intent (requires using RStudio)
}
