% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_add.R
\name{topic_add}
\alias{topic_add}
\title{Add your own topic to the existing topic list}
\usage{
topic_add(topics.current, topic.new)
}
\arguments{
\item{topics.current}{character vector.  Existing vector of topics (defaults to a vector called "topics" )}

\item{topic.new}{character.  New topic to add}
}
\value{
nothing.  updates existing vector.
}
\description{
This function updates an existing character vector (defaults a vector called'topics')
}
