% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_generator.R
\name{topic_generator}
\alias{topic_generator}
\title{Suggest random topic for discussion}
\usage{
topic_generator(topic_vector, num_topics = 1)
}
\arguments{
\item{topic_vector}{a vector of suggests topics}

\item{num_topics}{a number of topics to suggest, defaults to 1}
}
\value{
a vector of length one
}
\description{
This function returns one item from a vector, in this case a random topic
}
