% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\docType{data}
\name{topics}
\alias{topics}
\title{Fika topics.}
\format{character vector
\describe{
\item{topics}{character vector}
}}
\usage{
topics
}
\description{
List of potential fika topics
}
\examples{
  topics
}
\keyword{datasets}
